+++ pywren_ibm_cloud/function/invoker.py
@@ -118,7 +118,7 @@ class FunctionInvoker:
                    'executor_id': job.executor_id,
                    'job_id': job.job_id,
                    'call_id': call_id,
-                   'host_submit_time': time.time(),
+                   'host_submit_tstamp': time.time(),
                    'pywren_version': __version__,
                    'runtime_name': job.runtime_name,
                    'runtime_memory': job.runtime_memory}