+++ pywren_ibm_cloud/future.py
@@ -56,6 +56,7 @@ class ResponseFuture:
         self.runtime_name = job_description['runtime_name']
         self.runtime_memory = job_description['runtime_memory']
         self.activation_id = None
+        self.stats = {}
 
         self._storage_config = storage_config
         self._produce_output = True
@@ -68,10 +69,12 @@ class ResponseFuture:
         self._traceback = None
         self._call_status = None
         self._call_output = None
-        self._call_metadata = job_metadata.copy()
-
         self._status_query_count = 0
         self._output_query_count = 0
+
+        for key in job_metadata:
+            if any(ss in key for ss in ['time', 'tstamp', 'count', 'size']):
+                self.stats[key] = job_metadata[key]
 
         self._storage_path = get_storage_path(self._storage_config)
 
@@ -152,6 +155,8 @@ class ResponseFuture:
                 self._call_status = internal_storage.get_call_status(self.executor_id, self.job_id, self.call_id)
                 self._status_query_count += 1
 
+        self.stats['status_done_tstamp'] = time.time()
+        self.stats['status_query_count'] = self._status_query_count
         self.activation_id = self._call_status.pop('activation_id', None)
 
         if self._call_status['type'] == '__init__':
@@ -200,11 +205,12 @@ class ResponseFuture:
                 logger.debug('Exception: {} - {}'.format(self._exception[0].__name__, self._exception[1]))
                 return None
 
-        self._call_metadata['host_submit_time'] = self._call_status.pop('host_submit_time')
-        self._call_metadata['status_done_time'] = time.time()
-        self._call_metadata['status_query_count'] = self._status_query_count
+        for key in self._call_status:
+            if any(ss in key for ss in ['time', 'tstamp', 'count', 'size']):
+                self.stats[key] = self._call_status[key]
 
-        total_time = format(round(self._call_status['end_time'] - self._call_status['start_time'], 2), '.2f')
+        self.stats['exec_time'] = round(self.stats['end_tstamp'] - self.stats['start_tstamp'], 8)
+        total_time = format(round(self.stats['exec_time'], 2), '.2f')
 
         log_msg = ('ExecutorID {} | JobID {} - Got status from call {} - Activation '
                    'ID: {} - Time: {} seconds'.format(self.executor_id,
@@ -275,22 +281,20 @@ class ResponseFuture:
                 return None
 
         self._call_output = pickle.loads(call_output)
+        function_result = self._call_output['result']
 
-        self._call_metadata['output_done_time'] = time.time()
-        self._call_metadata['output_query_count'] = self._output_query_count
+        self.stats['output_done_tstamp'] = time.time()
+        self.stats['output_query_count'] = self._output_query_count
 
         log_msg = ('ExecutorID {} | JobID {} - Got output from call {} - Activation '
                    'ID: {}'.format(self.executor_id, self.job_id, self.call_id, self.activation_id))
         logger.info(log_msg)
 
-        function_result = self._call_output['result']
-
         if isinstance(function_result, ResponseFuture) or \
            (type(function_result) == list and len(function_result) > 0 and isinstance(function_result[0], ResponseFuture)):
             self._new_futures = [function_result] if type(function_result) == ResponseFuture else function_result
             self._set_state(ResponseFuture.State.Futures)
-            self._call_metadata['status_done_time'] = self._call_metadata['output_done_time']
-            del self._call_metadata['output_done_time']
+            self.stats['status_done_tstamp'] = self.stats.pop('output_done_tstamp')
             return self._new_futures
 
         else: