+++ pywren_ibm_cloud/compute/backends/docker/config.py
@@ -4,10 +4,10 @@ import tempfile
 from pywren_ibm_cloud.utils import version_str
 from pywren_ibm_cloud.config import LOGS_PREFIX
 
-RUNTIME_DEFAULT = {'3.5': 'ibmfunctions/pywren:3.5',
-                   '3.6': 'ibmfunctions/action-python-v3.6',
+RUNTIME_DEFAULT = {'3.5': 'ibmfunctions/pywren:3.5:latest',
+                   '3.6': 'ibmfunctions/action-python-v3.6:latest',
                    '3.7': 'ibmfunctions/action-python-v3.7:1.6.0',
-                   '3.8': 'jsampe/action-python-v3.8'}
+                   '3.8': 'jsampe/action-python-v3.8:latest'}
 RUNTIME_TIMEOUT_DEFAULT = 600  # 10 minutes
 
 PYWREN_SERVER_PORT = 8080
@@ -47,3 +47,6 @@ def load_config(config_data):
 
     if 'workers' not in config_data['pywren']:
         config_data['pywren']['workers'] = None
+
+    if 'ibm_cos' in config_data and 'private_endpoint' in config_data['ibm_cos']:
+        del config_data['ibm_cos']['private_endpoint']