+++ CHANGELOG.md
@@ -9,10 +9,11 @@
 
 ### Changed
 - [PyWren] Moved all stats from 'f._call_status' to a new 'f.stats' variable
+- [Localhost] Improved localhost storage backend
 
 
 ### Fixed
-- [PyWren] Fixd issue in pw.clean(cs=cobjs) when passing a large list of cobjs
+- [PyWren] Fixed issue in pw.clean(cs=cobjs) when passing a large list of cobjs
 
 
 ## [v1.5.2]